requirements.txt
numpy
pandas
astropy
scipy


flake8
pytest

pyyaml

pyqt5
pyqt5-tools
loguru
